requirements.txt
paho-mqtt==2.1.0
RPi.GPIO==0.7.1
smbus==1.1.post2
